refactor: Добавить index.ts с imoprt / export
Были добавлены index.ts для более удобной навигации по компонентам
