+++ src/components/app/index.ts
@@ -1,0 +1,2 @@
+import App from './app';
+export default App;
