+++ src/components/footer/index.ts
@@ -1,0 +1,2 @@
+import Footer from './footer';
+export default Footer;
