+++ src/components/header/index.ts
@@ -1,0 +1,2 @@
+import Header from './header';
+export default Header;
