+++ src/components/main-page-nav-location/index.ts
@@ -1,0 +1,3 @@
+import MainPageNavLocation from './main-page-nav-location';
+
+export default MainPageNavLocation;
