+++ src/components/map/index.ts
@@ -1,0 +1,2 @@
+import Map from './map';
+export default Map;
