+++ src/components/offer-card/index.ts
@@ -1,0 +1,3 @@
+import OfferList from './offer-list';
+
+export default OfferList;
