+++ src/components/private/index.ts
@@ -1,0 +1,2 @@
+import PrivateRoute from './private';
+export default PrivateRoute;
