+++ src/components/sort/index.ts
@@ -1,0 +1,2 @@
+import Sort from './sort';
+export default Sort;
