+++ src/pages/favorites-page/favorites-page-empty.tsx
@@ -1,5 +1,5 @@
-import Header from '../../components/header/header';
-import Footer from '../../components/footer/footer';
+import Header from '../../components/header';
+import Footer from '../../components/footer';
 
 function FavoritesPageEmpty(): JSX.Element {
   return (
