+++ src/pages/login-page/login-page.tsx
@@ -1,5 +1,5 @@
-import Footer from '../../components/footer/footer';
-import Header from '../../components/header/header';
+import Footer from '../../components/footer';
+import Header from '../../components/header';
 import { Link } from 'react-router-dom';
 import { APP_ROUTE } from '../../const';
